+++ hulusubs_dl/cust_utils/utils.py
@@ -129,3 +129,9 @@ def read_file_data(file_path, file_name):
     with open(file_location, 'r') as f:
         content = f.read().strip()
     return None if content == "" else content
+
+
+def get_clean_path_name(path_name):
+    remove_punctuation_map = dict((ord(char), None) for char in '\/*?:"<>|,;\'')
+    path_name.translate(remove_punctuation_map)
+    return path_name