+++ Changelog.md
@@ -1,3 +1,6 @@
+
 # Changelog
 - Re-did the entire codebase and implemented hulu API to scrap data.
 - Proxy support added.
+- Fix for #22
+- Added Verbose Logging.