+++ docs/Changelog.md
@@ -2,3 +2,5 @@
 # Changelog
 - Re-did the entire codebase and implemented hulu API to scrap data.
 - Proxy support added.
+- Fix for #22
+- Added Verbose Logging.